+++ src/zarr/zarr_chunk_layout.c
@@ -22,7 +22,7 @@ zarr_chunk_layout_probe(struct store* s,
                         uint64_t first_chunk_off,
                         uint32_t first_chunk_cbytes,
                         uint8_t codec_id,
-                        uint16_t max_substreams_per_chunk,
+                        uint32_t max_substreams_per_chunk,
                         struct chunk_layout* out)
 {
   CHECK_SILENT(Fail, s);
